requirements.txt
google-api-python-client==2.45.0

protobuf~=5.28.2